fix(product): close support diagnostics privacy gaps
