+++ apps/packages/product-client/src/lib/domain/support/report-upload-sanitizer.ts
@@ -1,41 +1,29 @@
-const SECRET_KEY_PATTERN = /(token|key|secret|password|authorization|credential|cookie)/i;
 const BEARER_PATTERN = /\b(bearer\s+)[A-Za-z0-9._\-/+=]+/gi;
-const LONG_TOKEN_PATTERN = /\b[A-Za-z0-9_\-+/=]{48,}\b/g;
+const ENV_SECRET_PATTERN =
+  /\b([A-Z0-9_]*(?:TOKEN|KEY|SECRET|PASSWORD|CREDENTIAL|COOKIE)[A-Z0-9_]*\s*[:=]\s*)(?:"[^"\r\n]*"|'[^'\r\n]*'|[^\s"',;]+)/gi;
+const LONG_TOKEN_PATTERN =
+  /(^|[^A-Za-z0-9_\-+/=])([A-Za-z0-9_\-+/=]{48,})(?=$|[^A-Za-z0-9_\-+/=])/g;
 const SIGNED_URL_PATTERN =
   /([?&](?:x-amz-signature|x-amz-credential|x-amz-security-token|signature|sig|token|key|secret)=)[^&#\s]+/gi;
+const MAX_LOG_TEXT_CHARS = 2 * 1024 * 1024;
 
-export function sanitizeSupportUploadPayload<T>(value: T): T {
-  return sanitizeValue(value) as T;
-}
-
-function sanitizeValue(value: unknown, keyHint = ""): unknown {
-  if (value == null) {
-    return value;
+export function sanitizeSupportLogText(value: unknown): string {
+  if (typeof value !== "string") {
+    return "[redacted]";
   }
-  if (typeof value === "string") {
-    return SECRET_KEY_PATTERN.test(keyHint) ? "[REDACTED]" : sanitizeString(value);
-  }
-  if (typeof value !== "object") {
-    return value;
-  }
-  if (Array.isArray(value)) {
-    return value.map((item) => sanitizeValue(item, keyHint));
-  }
-
-  const output: Record<string, unknown> = {};
-  for (const [key, item] of Object.entries(value)) {
-    if (SECRET_KEY_PATTERN.test(key)) {
-      output[key] = "[REDACTED]";
-      continue;
-    }
-    output[key] = sanitizeValue(item, key);
-  }
-  return output;
-}
-
-function sanitizeString(value: string): string {
-  return value
+  const boundedValue = boundedLogTail(value);
+  return boundedValue
     .replace(BEARER_PATTERN, "$1[REDACTED]")
+    .replace(ENV_SECRET_PATTERN, "$1[REDACTED]")
     .replace(SIGNED_URL_PATTERN, "$1[REDACTED]")
-    .replace(LONG_TOKEN_PATTERN, "[REDACTED]");
+    .replace(LONG_TOKEN_PATTERN, "$1[REDACTED]");
+}
+
+function boundedLogTail(value: string): string {
+  if (value.length <= MAX_LOG_TEXT_CHARS) {
+    return value;
+  }
+  const tail = String.prototype.slice.call(value, -MAX_LOG_TEXT_CHARS) as string;
+  const firstLineEnd = tail.indexOf("\n");
+  return firstLineEnd >= 0 ? tail.slice(firstLineEnd + 1) : "";
 }
