+++ specs/codebase/systems/product/support/README.md
@@ -173,9 +173,12 @@ The schema-v2 package omits the later active-workspace/session fields. Structure
 identifiers, customer-provided names/paths, and runtime URLs are represented only
 by length placeholders; trusted correlation remains in the server-owned request
 record. Package assembly uses bounded, explicit projections; malformed scalar or
-container values fail closed before the final text scrub. Native collection failures use fixed
-`desktop`, `anyharness`, or generic diagnostic classes and never include paths or
-operating-system error text.
+container values and revoked proxies fail closed. Timestamps, versions, runtime
+states, platforms, and attachment media types must match their audited field
+contracts. Only bounded diagnostic log text receives credential, signed-URL,
+and opaque-token scrubbing; ordinary diagnostic text remains useful. Native
+collection failures use fixed `desktop`, `anyharness`, or generic diagnostic
+classes and never include paths or operating-system error text.
 
 The package runs session data through the session-debug sanitizer. Prompt and
 message bodies, raw tool input/output, event content, notification bodies, and
